Remove postbuild script and replace with pretest
Having the tests run on `postbuild` - which is called by `prepublish`, and hence also run on `postinstall` - causes failing tests to cause `npm install` to fail. This means that if the tests are not currently passing then npm rolls back the install, which is fairly hostile for both CI and development.

Instead invert the causality so that tests only run when explicitly called.