Allow for totally custom proposition headers
If `propositionHeaderLink` is not set (specifically if it is set to a falsy value to avoid breaking change) but a `propositionHeader` _is_ set, then render the `propositionHeader` without wrapping it in a link.

This is to allow fully custom header content - specifically for my use case a "logged in user" state.